♻️ Extract components (#967)
