+++ src/lib/components/TaskList.svelte
@@ -10,13 +10,13 @@
     TableHeadCell,
   } from 'flowbite-svelte';
 
-  import type { TaskResult, TaskResults } from '$lib/types/task';
-  import type { SubmissionRatios } from '$lib/types/submission';
+  import type { TaskResults } from '$lib/types/task';
 
   import ThermometerProgressBar from '$lib/components/ThermometerProgressBar.svelte';
   import UpdatingModal from '$lib/components/SubmissionStatus/UpdatingModal.svelte';
   import SubmissionStatusImage from '$lib/components/SubmissionStatus/SubmissionStatusImage.svelte';
-  import { getBackgroundColorFrom, submission_statuses } from '$lib/services/submission_status';
+  import AcceptedCounter from '$lib/components/SubmissionStatus/AcceptedCounter.svelte';
+  import { getBackgroundColorFrom } from '$lib/services/submission_status';
   import { ATCODER_BASE_CONTEST_URL } from '$lib/constants/urls';
   import { getContestNameLabel } from '$lib/utils/contest';
   import { taskUrl, toWhiteTextIfNeeds } from '$lib/utils/task';
@@ -26,37 +26,6 @@
   export let taskResults: TaskResults;
   export let isAdmin: boolean;
   export let isLoggedIn: boolean;
-
-  // TODO: ユーザの設定に応じて、ACかどうかの判定を変更できるようにする
-  // TODO: 別ファイルに切り出す
-  let acceptedCount: number;
-  let acceptedRatioPercent: number;
-  let submissionRatios: SubmissionRatios;
-
-  const getRatioPercent = (taskResults: TaskResults, statusName: string) => {
-    const count = taskResults.filter((taskResult) => taskResult.status_name === statusName).length;
-    const ratioPercent = (count / taskResults.length) * 100;
-
-    return ratioPercent;
-  };
-
-  $: {
-    acceptedCount = taskResults.filter((taskResult: TaskResult) => taskResult.is_ac).length;
-    acceptedRatioPercent = (acceptedCount / taskResults.length) * 100;
-
-    submissionRatios = submission_statuses
-      .filter((status) => status.status_name !== 'ns')
-      .map((status) => {
-        const name = status.status_name;
-        const results = {
-          name: name,
-          ratioPercent: getRatioPercent(taskResults, name),
-          color: status.background_color,
-        };
-
-        return results;
-      });
-  }
 
   let updatingModal: UpdatingModal;
 </script>
@@ -68,18 +37,8 @@
         {grade}
       </div>
 
-      <ThermometerProgressBar {submissionRatios} />
-
-      <!-- See: -->
-      <!-- https://developer.mozilla.org/ja/docs/Web/JavaScript/Reference/Global_Objects/Number/toFixed -->
-      <div class="text-sm w-1/12 text-center">
-        <div>
-          {acceptedCount} / {taskResults.length}
-        </div>
-        <div>
-          {`(${acceptedRatioPercent.toFixed(1)}%)`}
-        </div>
-      </div>
+      <ThermometerProgressBar {taskResults} />
+      <AcceptedCounter {taskResults} />
     </span>
 
     <!-- FIXME: clickを1回実行するとactionsが2回実行されてしまう。原因と修正方法が分かっていない。 -->
